refactor: extract OMP command/lifecycle and version-pin modules
Splits src/harness.rs into src/harness/{mod,omp,version_pin}.rs ahead of
Phase 1 seat execution, per docs/plans/productization-2026-07-17.md and
Powder cerberus-043 criterion 1. No behavior change: same 138 library
tests pass, same test count, same public API surface (OmpSubstrateConfig
re-exported at crate::harness::OmpSubstrateConfig unchanged).
